Testing handling of prematurely stopped loop.
Testing when the service is stopped but unable to clean up properly
before shutting down. Really testing that the service doesn't hang when
user attempts to stop it.

Found a bug where if ffmpeg refuses to die after SIGKILL, the service
would get unhandled exceptions since the last item in the
timeout_signals list was popped. Using an infinite generator of SIGKILLS
to resolve the bug.